# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateReservationRequest": {
        "description": "The host is the signed-in user; nothing here names them.",
        "properties": {
          "capacity": {
            "default": 1,
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Capacity",
            "type": "integer"
          },
          "match_type": {
            "$ref": "#/components/schemas/MatchType"
          },
          "memo": {
            "default": "",
            "maxLength": 140,
            "title": "Memo",
            "type": "string"
          },
          "ranks": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Ranks",
            "type": "array"
          },
          "start_time": {
            "format": "time",
            "title": "Start Time",
            "type": "string"
          }
        },
        "required": [
          "start_time",
          "match_type"
        ],
        "title": "CreateReservationRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JoinReservationRequest": {
        "properties": {
          "ranks": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Ranks",
            "type": "array"
          }
        },
        "title": "JoinReservationRequest",
        "type": "object"
      },
...
      "MatchType": {
        "enum": [
          "rank_match",
          "player_match",
          "any"
        ],
        "title": "MatchType",
        "type": "string"
      },
      "ParticipantSummaryOut": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "id",
          "display_name",
          "username"
        ],
        "title": "ParticipantSummaryOut",
        "type": "object"
      },
...
      "ReservationOut": {
        "properties": {
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "host_display_name": {
            "title": "Host Display Name",
            "type": "string"
          },
          "host_ranks": {
            "items": {
              "type": "string"
            },
            "title": "Host Ranks",
            "type": "array"
          },
          "host_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host Username"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "match_type": {
            "$ref": "#/components/schemas/MatchType"
          },
          "memo": {
            "title": "Memo",
            "type": "string"
          },
          "participant_count": {
            "title": "Participant Count",
            "type": "integer"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/ParticipantSummaryOut"
            },
            "title": "Participants",
            "type": "array"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReservationStatus"
          }
        },
        "required": [
          "id",
          "start_at",
          "host_display_name",
          "host_username",
          "host_ranks",
          "match_type",
          "capacity",
          "memo",
          "status",
          "participant_count",
          "participants",
          "created_at"
        ],
        "title": "ReservationOut",
        "type": "object"
      },
      "ReservationStatus": {
        "enum": [
          "open",
          "matched",
          "cancelled",
          "ended"
        ],
        "title": "ReservationStatus",
        "type": "string"
      },
...
      "UpdateReservationRequest": {
        "description": "Every field optional: the host sends only what changed.\n\nThe validators are the create request's own, so a rule cannot drift between\nposting a reservation and editing one.",
        "properties": {
          "capacity": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "match_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchType"
              },
              {
                "type": "null"
              }
            ]
          },
          "memo": {
            "anyOf": [
              {
                "maxLength": 140,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memo"
          },
          "ranks": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranks"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          }
        },
        "title": "UpdateReservationRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "reservation__models__CommentOut": {
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "author_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Username"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "reservation_id": {
            "title": "Reservation Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "reservation_id",
          "author",
          "author_username",
          "body",
          "created_at"
        ],
        "title": "CommentOut",
        "type": "object"
      },
      "reservation__models__CreateCommentRequest": {
        "description": "Whitespace is stripped before the length rules run.\n\nA body of three spaces is an empty comment; letting min_length see it\nunstripped would store one, because the service strips on the way to the\nrepository either way.",
        "properties": {
          "body": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "CreateCommentRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "`POST /auth/login`의 access_token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Live data from the RPCN multiplayer server for TTT2.",
    "title": "Tekken Tag Tournament 2 RPCN API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reservations": {
      "get": {
        "description": "Reservations from now until the next 06:00 KST.",
        "operationId": "list_reservations_reservations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReservationOut"
                  },
                  "title": "Response List Reservations Reservations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Reservations",
        "tags": [
          "Reservations"
        ]
      },
      "post": {
        "operationId": "create_reservation_reservations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Reservation",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{reservation_id}": {
      "delete": {
        "operationId": "cancel_reservation_reservations__reservation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "title": "Reservation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Reservation",
        "tags": [
          "Reservations"
        ]
      },
      "get": {
        "operationId": "get_reservation_reservations__reservation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "title": "Reservation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Reservation",
        "tags": [
          "Reservations"
        ]
      },
      "patch": {
        "operationId": "update_reservation_reservations__reservation_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "title": "Reservation Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Reservation",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{reservation_id}/comments": {
      "get": {
        "operationId": "list_comments_reservations__reservation_id__comments_get",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "title": "Reservation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/reservation__models__CommentOut"
                  },
                  "title": "Response List Comments Reservations  Reservation Id  Comments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Comments",
        "tags": [
          "Reservations"
        ]
      },
      "post": {
        "operationId": "add_comment_reservations__reservation_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "title": "Reservation Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/reservation__models__CreateCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/reservation__models__CommentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Comment",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{reservation_id}/comments/{comment_id}": {
      "delete": {
        "operationId": "delete_comment_reservations__reservation_id__comments__comment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "title": "Reservation Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Comment",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{reservation_id}/participants": {
      "post": {
        "operationId": "join_reservation_reservations__reservation_id__participants_post",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "title": "Reservation Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinReservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Join Reservation",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{reservation_id}/participants/me": {
      "delete": {
        "operationId": "cancel_participation_reservations__reservation_id__participants_me_delete",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "title": "Reservation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Participation",
        "tags": [
          "Reservations"
        ]
      }
    },
...
  }
}